Move router trie to its own file
Just as the title says.   Trie is useful outside of just the router (for
making other routers.)
